requirements.txt
python-nmap
dash
pandas
plotly